added reducers pages. Our state refactoring of the function reduce Profile and Dialogs
